backend/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
python-multipart
pydantic
pypdf